// www/js/GroupForums/ts/react/shared/types.ts
interface Reaction {
  emoteId: string;
  reactionCount: number;
  hasUserAppliedReaction: boolean;
  areReactionCountsVisible: boolean;
}